gemini_bridge/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
websockets>=11.0.3
google-generativeai>=0.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
